+++ run_tests.sh
@@ -1,6 +1,6 @@
 if [ ${1:-not_full} == "full" ]
 then
-    pytest -k "main" --log-level="INFO" --html=report.html --self-contained-html --record-mode=new_episodes -vv "${@:2}"
+    python -m pytest -k "main" --log-level="INFO" --html=report.html --self-contained-html --record-mode=new_episodes -vv "${@:2}"
 else
-    pytest -k "not main" --log-level="INFO" --html=report.html --self-contained-html --record-mode=new_episodes -vv "$@"
+    python -m pytest -k "not main and not test_client" --log-level="INFO" --html=report.html --self-contained-html --record-mode=new_episodes -vv "$@"
 fi